requirements.txt
requests
beautifulsoup4
openai
python-dotenv
tenacity
